requirements.txt
psycopg2-binary
python-dotenv
langchain
tiktoken
chromadb
python-dotenv
openai
